requirements.txt
streamlit
openai
python-dotenv
langchain>=0.2,<0.3

